Restore moving the default part down, after get_certificate_list()
If ask_pin is false, we try to do the login as early as possible (as
this may give access to the list in more cases; this may also invoke
the pinpad in some rare cases). Anyway the idea of moving it further
down was just to avoid early asking of the PIN, which is anyway
disabled if ask_pin is false.

If ask_pin is true, asking for PIN is delayed.